Update the build and launch scripts
Update the build script to accept either linux or kernel for building the
kernel support and support for patching ovmf and qemu (kernel patching is
already supported).

Update the launch script for when a system is using bridged networking.
Currently part of the MAC address is using a hardcoded values in the MAC
address suffix which can lead to MAC address conflicts when run on more
than one machine in a network. Use a portion of the bridge MAC address
when generating the guest MAC address to try to minimize this conflict.