Remove virtio-rng-pci from the launch script
Using 'virtio-rng-pci' would register a device controlled by the HV
as a source of entropy. Since the SEV attacker model considers the HV
as untrusted, the Guest should not rely on the HV to add entropy to
its entropy pool.
This shouldn't necessarily be a security risk since the Guest receives
entropy from other sources as well but it would be good to address this.
The device must have been added to the script by mistake consindering
the comment discusses the virtio ring while the device name is virtio
rng.