Disable virtio-rng and virtio-crypto for the guest kernel
VirtioRNG allows the host to inject random data into the /dev/hwrng
device in the guest which is undesirable. Thus, we disable it in the
guest's build script.
Using VirtioCrypto would expose AES cryptographic operations, both keys
and data, so we disable it as well.
This patch disables them also for the host in case this build script is
used for the host kernel. However, this should not be an issue since
the virtio drivers are only used by the guest.