launch-qemu.sh: don't enable usermode networking by default
As of QEMU 7.2 usermode networking requires a fairly recent libslirp
that is likely to not be available during the build. This will cause
launch-qemu.sh to always fail by default on such a host. Instead, add
a new -default-networking option and only enable it when the user
specifically specifies it.

Signed-off-by: Michael Roth <michael.roth@amd.com>