Enable additional kernel config options by default
CONFIG_MODULE_SIG_KEY seems to be a common source of reported build
breakages in Ubuntu 22.04 build environments, so disable it by default.

CONFIG_X86_CPUID has always been required since launch-qemu.sh relies
on it to determine C-bit position, and with newer distros don't enable
it by default so it needs to be done explicitly.

Signed-off-by: Michael Roth <michael.roth@amd.com>