Increase maximums for vCPUs/memory
The current hard-code limits of 64 vCPUs and 30G of memory are too
restrictive for testing on larger systems. Set vCPU max to the 255 limit
dictated by use of APIC (since x2APIC not currently supported) and make
the memory max equal to whatever the user specifies + 8GB to allow for
memory hotplug testing (not tested but may be possible in conjunction
with guests kernels that support lazy acceptance).

Signed-off-by: Michael Roth <michael.roth@amd.com>